+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# Alpha v0.2
-* When outside a rested area, the HUD will now show while town clothes are equipped.
+# Alpha v0.3
+* Improved initial load flow after loading in to reduce issues creating equipment sets.
+
+# Alpha v0.2.1
+* When outside a rested area, the HUD will now show while town clothes are equipped. 
 
 # Alpha v0.1
 * Initial alpha release.
+